moved NameField to new fields folder
