+++ components/forms/SignUpForm.jsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useState } from "react";
 import { cn } from "@/lib/utils";
-import { NameField } from "../ui/nameField";
+import { NameField } from "../ui/Fields/nameField";
 import { EmailField } from "../ui/Fields/emailField";
 import { Button } from "../ui/button";
 import { SubmitButton } from "../ui/submitButton";
